reservations_backend/pyproject.toml
[project]
name = "reservation-fullstack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-cors-headers>=4.6.0",
    "django-phonenumber-field>=8.0.0",
    "django-rest-framework>=0.1.0",
    "django>=5.1.1",
    "djangorestframework-simplejwt>=5.3.1",
    "drf-spectacular>=0.27.2",
    "phonenumbers>=8.13.46",
    "pillow>=10.4.0",
]

[tool.uv]
dev-dependencies = [
    "bandit>=1.7.10",
]
